requirements.txt
-e .
aws-cdk.aws-ec2